requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.0      
django-allauth>=0.54      
django-cors-headers>=4.0   
Pillow>=10.0               
psycopg2-binary>=2.9       
gunicorn>=21.2             
python-dotenv>=1.0         